Replace setup/i18n.py and ui/gtk2/i18n.py
setup/i18n.py was a symlink of ui/gtk2/i18n.py but ui/gtk2 is deprecated
so replace the two files.
